+++ mogo-core/src/main/java/io/github/loserya/module/interceptor/tenantline/TenantLineInterceptor.java
@@ -109,7 +109,7 @@ public abstract class TenantLineInterceptor implements Interceptor {
     }
 
     @Override
-    public final Object[] page(LambdaQueryWrapper<?> queryWrapper, int pageNo, int pageSize, Class<?> clazz) {
+    public final Object[] page(LambdaQueryWrapper<?> queryWrapper, Long pageNo, Long pageSize, Class<?> clazz) {
         addTenantLineCondition(queryWrapper, clazz);
         return Interceptor.super.page(queryWrapper, pageNo, pageSize, clazz);
     }
